Follows proper autoloading standards  - Splits PHP classes in descriptor.php into separate files  - Splits MapFieldIter and RepeatedFieldIter into separate files  - Moves descriptor.php to Internal/functions.php  - Moves all namespaced functions into Iternal/functions.php
